add section for start time field
